add efficient display weights from first layer
