+++ src/optimization/hand_tuned.py
@@ -71,8 +71,8 @@ EXPERIMENTS = [
         train_test_directories=["data/aptos2019-blindness-detection/train_images"],
         model=("efficientnet-b0", {"num_classes": 5, "pretrained": True}),
         batch_size=58,
-        optimizer=("Adam", {"lr": 1e-5 , "weight_decay": 0.001}),
-        test_size=0.4,
+        optimizer=("Adam", {"lr": 1e-5 , "weight_decay": 0.0001}),
+        test_size=0.2,
         max_epochs=120,
         sampler=("RandomSampler", {}),
         lr_scheduler=("ExponentialLR", {"gamma": 0.99})
